refactor(conversion): replace toy model generator
